add entities upserted
