+++ packages/datatrak-web/src/api/mutations/submitSurvey/useSubmitSurvey.ts
@@ -1,0 +1,59 @@
+/*
+ * Tupaia
+ *  Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
+ */
+
+import { useMutation } from 'react-query';
+import { generatePath, useNavigate, useParams } from 'react-router';
+import { Coconut } from '../../../components';
+import { post } from '../../api';
+import { ROUTES } from '../../../constants';
+import { useSurveyForm } from '../../../features';
+import { useSurvey, useUser } from '../../queries';
+import { successToast } from '../../../utils';
+import { processSurveyResponse } from './processSurveyResponse';
+import { Answers } from './types';
+
+// utility hook for getting survey response data
+export const useSurveyResponseData = () => {
+  const { data: user } = useUser();
+  const { surveyCode } = useParams();
+  const { surveyStartTime, surveyScreenComponents } = useSurveyForm();
+  const { data: survey } = useSurvey(surveyCode);
+  return {
+    surveyStartTime,
+    surveyId: survey?.id,
+    questions: Object.values(surveyScreenComponents!).reduce((acc, val) => acc.concat(val), []), // flattened array of survey questions
+    countryId: user?.country?.id,
+  };
+};
+
+export const useSubmitSurvey = () => {
+  const navigate = useNavigate();
+  const params = useParams();
+
+  const surveyResponseData = useSurveyResponseData();
+
+  return useMutation<any, Error, Answers, unknown>(
+    async (answers: Answers) => {
+      if (!answers) {
+        return;
+      }
+
+      const processedResponse = processSurveyResponse({
+        ...surveyResponseData,
+        answers,
+      });
+
+      await post('submitSurvey', {
+        data: processedResponse,
+      });
+    },
+    {
+      onSuccess: () => {
+        successToast("Congratulations! You've earned a coconut", Coconut);
+        navigate(generatePath(ROUTES.SURVEY_SUCCESS, params));
+      },
+    },
+  );
+};
