+++ packages/datatrak-web/src/api/mutations/submitSurvey/utils.ts
@@ -1,0 +1,14 @@
+/*
+ * Tupaia
+ *  Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
+ */
+
+export const isUpsertEntityQuestion = config => {
+  if (!config?.entity) {
+    return false;
+  }
+  if (config.entity.createNew) {
+    return true;
+  }
+  return config.entity.fields && Object.keys(config.entity.fields).length > 0;
+};
