+++ packages/datatrak-web/src/components/TextInput.tsx
@@ -29,22 +29,17 @@ interface TextInputProps
   ref?: React.Ref<HTMLInputElement>;
   textInputProps?: TextFieldProps;
 }
-export const TextInput = ({
-  value,
-  label,
-  name,
-  onChange,
-  id,
-  ref,
-  textInputProps,
-}: TextInputProps) => (
-  <Label
-    label={label}
-    name={name}
-    inputRef={ref}
-    labelPlacement={'top'}
-    onChange={onChange}
-    value={value}
-    control={<TextField id={id} {...textInputProps} fullWidth />}
-  />
-);
+export const TextInput = React.forwardRef<HTMLDivElement, TextInputProps>((props, ref) => {
+  const { value, label, name, onChange, id, textInputProps } = props;
+  return (
+    <Label
+      label={label}
+      name={name}
+      inputRef={ref}
+      labelPlacement={'top'}
+      onChange={onChange}
+      value={value}
+      control={<TextField id={id} {...textInputProps} fullWidth />}
+    />
+  );
+});
