--- packages/datatrak-web/src/features/Survey/SurveySideMenu/index.ts
@@ -1,6 +1,0 @@
-/*
- * Tupaia
- *  Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
- */
-
-export { SurveySideMenu, SIDE_MENU_WIDTH } from './SurveySideMenu';
