+++ packages/datatrak-web/src/types/surveys.ts
@@ -2,7 +2,6 @@
  * Tupaia
  *  Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
  */
-import { Ref } from 'react';
 import { ControllerRenderProps } from 'react-hook-form';
 import { DatatrakWebSurveysRequest, DatatrakWebSurveyScreenComponentsRequest } from '@tupaia/types';
